menus controller update
