--- app/helpers/menu_helper.rb
@@ -1,2 +1,0 @@
-module MenuHelper
-end
