+++ app/helpers/menus_helper.rb
@@ -1,0 +1,2 @@
+module MenusHelper
+end
