--- app/views/menu/destroy.html.erb
@@ -1,2 +1,0 @@
-<h1>Menu#destroy</h1>
-<p>Find me in app/views/menu/destroy.html.erb</p>
