--- app/views/menu/edit.html.erb
@@ -1,2 +1,0 @@
-<h1>Menu#edit</h1>
-<p>Find me in app/views/menu/edit.html.erb</p>
