--- app/views/menu/index.html.erb
@@ -1,4 +1,0 @@
-<h1>Menu#index</h1>
-<p>Find me in app/views/menu/index.html.erb</p>
-
-<h4> Welcome to you profile page <strong><i><%= current_user.username %></i></strong> </h4>
