--- app/views/menu/new.html.erb
@@ -1,4 +1,0 @@
-<h1>Menu#new</h1>
-<p>Find me in app/views/menu/new.html.erb</p>
-
-<%= render 'form', profile: @menu %>
