--- app/views/menu/update.html.erb
@@ -1,2 +1,0 @@
-<h1>Menu#update</h1>
-<p>Find me in app/views/menu/update.html.erb</p>
