+++ app/views/menus/destroy.html.erb
@@ -1,0 +1,2 @@
+<h1>Menus#destroy</h1>
+<p>Find me in app/views/menus/destroy.html.erb</p>
