+++ app/views/menus/edit.html.erb
@@ -1,0 +1,2 @@
+<h1>Menus#edit</h1>
+<p>Find me in app/views/menus/edit.html.erb</p>
