+++ app/views/menus/index.html.erb
@@ -1,0 +1,2 @@
+<h1>Menus#index</h1>
+<p>Find me in app/views/menus/index.html.erb</p>
