+++ app/views/menus/new.html.erb
@@ -1,0 +1,5 @@
+<h1>Menus#new</h1>
+<p>Find me in app/views/menus/new.html.erb</p>
+
+
+<%= render 'form', profile: @menu %>
