+++ app/views/menus/show.html.erb
@@ -1,0 +1,2 @@
+<h1>Menus#show</h1>
+<p>Find me in app/views/menus/show.html.erb</p>
