+++ app/views/menus/update.html.erb
@@ -1,0 +1,2 @@
+<h1>Menus#update</h1>
+<p>Find me in app/views/menus/update.html.erb</p>
